Feature spring controller

+++ src/main/java/ua/su/service/impl/ClinicServiceImpl.java
@@ -29,7 +29,7 @@ public class ClinicServiceImpl implements ClinicService {
     public Clinic getEntry(Long id) {
         Clinic clinic = clinicRepository.getOne(id);
         if (clinic == null) {
-            throw new RuntimeException("Cannot find cat with id " + id);
+            throw new RuntimeException("Cannot find clinic with id " + id);
         }
         List<MedicalProcedure> medicalProcedures = clinicRepository.getAllByClinicId(id);
         clinic.setMedicalProcedures(medicalProcedures);
@@ -38,7 +38,10 @@ public class ClinicServiceImpl implements ClinicService {
 
     @Override
     public void delete(Long id) {
-        clinicRepository.getOne(id);
+        Clinic clinic = clinicRepository.getOne(id);
+        if (clinic == null) {
+            throw new RuntimeException("Cannot find clinic with id " + id);
+        }
         medicalProcedureRepository.deleteByClinicId(id);
         clinicRepository.delete(id);
     }
@@ -59,6 +62,10 @@ public class ClinicServiceImpl implements ClinicService {
     @Override
     public Clinic update(Long id, Clinic clinic) {
         clinicRepository.getOne(id);
+        Clinic clinicFromRepository = clinicRepository.getOne(id);
+        if (clinicFromRepository == null) {
+            throw new RuntimeException("Cannot find clinic with id " + id);
+        }
         medicalProcedureRepository.deleteByClinicId(id);
         clinicRepository.update(id, clinic);
         List<MedicalProcedure> mpList = clinic.getMedicalProcedures();
